+++ src/app/commands/scoap-squad/ScoapSquad.ts
@@ -89,7 +89,7 @@ module.exports = class ScoapSquad extends SlashCommand {
 		command.catch(e => {
 			if (!(e instanceof ValidationError)) {
 				LogUtils.logError('failed to handle scoap-squad command', e);
-				return ctx.send('Sorry something is not working and our devs are looking into it');
+				return ServiceUtils.sendOutErrorMessage(ctx);
 			}
 		});
 	}