+++ src/app/commands/timecard/Timecard.ts
@@ -83,10 +83,10 @@ export default class Timecard extends SlashCommand {
 			return ctx.send(`${ctx.user.mention} Sent you a DM with information.`);
 		}).catch(e => {
 			if (e instanceof ValidationError) {
-				return ctx.send(e.message);
+				return ctx.send({ content: `${e.message}`, ephemeral: true });
 			} else {
 				LogUtils.logError('failed to handle timecard command', e);
-				return ctx.send('Sorry something is not working and our devs are looking into it.');
+				return ServiceUtils.sendOutErrorMessage(ctx);
 			}
 		});
 	}