+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## 1.0.1-RELEASE (20201-11-24)
+## 1.1.0-RELEASE (2021-11-26)
+
+1. Handle edge cases for first quest
+   - flow captcha in text channel only
+   - add first quest start command
+   - enhance stability
+
+## 1.0.1-RELEASE (2021-11-24)
 
 1. Add more phrases and add some logging to first quest
 2. Generate new captcha on failure